+++ Riot/Modules/DeviceVerification/Verify/DeviceVerificationVerifyViewController.swift
@@ -78,18 +78,6 @@ final class DeviceVerificationVerifyViewController: UIViewController {
         // Hide back button
         self.navigationItem.setHidesBackButton(true, animated: animated)
     }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-    }
-    
-    override func viewDidDisappear(_ animated: Bool) {
-        super.viewDidDisappear(animated)
-    }
-    
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-    }
     
     override var preferredStatusBarStyle: UIStatusBarStyle {
         return self.theme.statusBarStyle