+++ CHANGES.rst
@@ -12,6 +12,9 @@ Improvements:
 Bug fix:
  * Registration with an email is broken (#2417).
 
+Bug fix:
+ * Device Verification: Fix user display name and device id colors in dark theme
+
 Changes in 0.8.6 (2019-05-06)
 ===============================================